[usmon 690]: Infrastructure to allow tls raw payload tests (#22036)
* usm: http2: tests: Add empty file for migrating future tests

* usm: http2: tests: Migrated raw payload tests to a new file

The migration is a preliminary step in order to allow running the same tests
both for plain text and tls versions.

* usm: http2: tests: enable gotls is we're in TLS mode

* usm: http2: tests: Replace monitor with tracer

The monitor cannot capture TLS traffic on its own - so we need the tracer to exist instead

* usm: http2: tests: Make changes to the consts

* usm: http2: Make a use of the proxy server
